pyproject.toml
[tool.poetry]
name = "hnac"
version = "0.7.0"
description = "Crawler for the Hackernews API"
authors = ["Matigakis Panagiotis <pmatigakis@gmail.com>"]
license = "MIT"
include = [
  "LICENSE",
  "README.md",
  "alembic.ini",
  "hnac/web/templates/**/*",
  "hnac/web/static/**/*",
]

[tool.poetry.dependencies]
python = "^3.8.0"
Flask = "1.1.2"
marshmallow = "3.8.0"
Flask-Script = "2.0.6"
Flask-Login = "0.5.0"
Flask-WTF = "0.14.3"
Flask-RESTful = "0.3.8"
alembic = "1.4.3"
psycopg2 = "2.8.6"
Flask-Admin = "1.5.6"
uWSGI = "2.0.19.1"
pika = "1.1.0"
arrow = "0.16.0"
Flask-SQLAlchemy = "2.4.4"
flask-uauth = "0.2.0"
python-dateutil = "2.8.1"
MarkupSafe = "1.1.1"
itsdangerous = "1.1.0"
Werkzeug = "1.0.1"
SQLAlchemy = "1.3.19"
python-dotenv = "0.20.0"

[tool.poetry.dev-dependencies]
httpretty = "^1.1.4"
nose = "^1.3.7"
responses = "^0.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
hnac-cli = "hnac.cli.cli:main"
